requirements.txt
pandas
scikit-learn
kafka-python
sqlalchemy
joblib
python-dotenv
